pyproject.toml
[project]
name = "lg-mentoring"
version = "2024.12.14"
description = "A video-based mentoring program."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3",
    "crispy-bootstrap5",
    "dj-database-url",
    "django-allauth",
    "django-browser-reload",
    "django-crispy-forms",
    "django-debug-toolbar",
    "django-embed-video",
    "django-filter",
    "django-storages[s3]",
    "django-upgrade",
    "django==5.1.4",
    "gunicorn",
    "pillow",
    "psycopg[binary]",
    "python-dotenv",
    "sentry-sdk",
    "whitenoise",
]
